move tests from test_summarize_new.py to test_summarize.py
