+++ skore/tests/unit/reports/estimator/metrics/test_summarize.py
@@ -1,3 +1,18 @@
+"""Tests for EstimatorReport.metrics.summarize().
+
+Organised by metric input type, then corner cases:
+
+- Default metrics — by ML task variant
+- Metric strings — skore built-ins and sklearn scorer names
+- Metric scorers — make_scorer / get_scorer instances
+- Metric callables — plain functions
+- Mixed metric types — strings + scorers + callables in one call
+- metric_kwargs
+- Cache and data_source
+"""
+
+import re
+
 import numpy as np
 import pandas as pd
 import pytest
@@ -11,7 +26,9 @@ from sklearn.metrics import (
     f1_score,
     get_scorer,
     make_scorer,
+    mean_absolute_error,
     precision_score,
+    r2_score,
     recall_score,
 )
 from sklearn.model_selection import train_test_split
@@ -27,20 +44,7 @@ def check_display_structure(
     expected_estimator_name=None,
     expected_data_source="test",
 ):
-    """
-    Helper function to check the structure of a MetricsSummaryDisplay.data DataFrame.
-
-    Parameters
-    ----------
-    display : MetricsSummaryDisplay
-        The display object to check.
-    expected_metrics : set, optional
-        Expected set of metric names.
-    expected_estimator_name : str, optional
-        Expected estimator name.
-    expected_data_source : str, default="test"
-        Expected data source value.
-    """
+    """Check the full structure of a MetricsSummaryDisplay.data DataFrame."""
     assert isinstance(display.data, pd.DataFrame)
     data = display.data
 
@@ -62,14 +66,79 @@ def check_display_structure(
     assert set(data["favorability"]) == {"(↗︎)", "(↘︎)"}
 
 
-# Test the happy path
+# ---------------------------------------------------------------------------
+# Default metrics
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.parametrize("metric", [None, [], {}])
+def test_default_metrics_plain(forest_binary_classification_with_test, metric):
+    """metric=None / [] / {} selects the canonical defaults for the ML task."""
+    estimator, X_test, y_test = forest_binary_classification_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    display = report.metrics.summarize(metric=metric)
+
+    check_display_structure(
+        display,
+        expected_metrics={
+            "Accuracy",
+            "Precision",
+            "Recall",
+            "ROC AUC",
+            "Log loss",
+            "Brier score",
+            "Fit time (s)",
+            "Predict time (s)",
+        },
+        expected_estimator_name="RandomForestClassifier",
+    )
+
+
+def test_default_metrics_with_extra_args(linear_regression_multioutput_with_test):
+    """metric=None with metric_kwargs alters how the default metrics are computed.
+
+    For multioutput regression, multioutput='raw_values' produces one row per
+    output for R² and RMSE.
+    """
+    estimator, X_test, y_test = linear_regression_multioutput_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    display = report.metrics.summarize(metric_kwargs={"multioutput": "raw_values"})
+
+    assert isinstance(display, MetricsSummaryDisplay)
+    r2_rows = display.data[display.data["metric"] == "R²"]
+    rmse_rows = display.data[display.data["metric"] == "RMSE"]
+    assert len(r2_rows) >= 2
+    assert len(rmse_rows) >= 2
+
+
+def test_default_metrics_as_dict(linear_regression_with_test):
+    """Built-in metric names as a dict: dict keys become the display names."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    display = report.metrics.summarize(metric={"My R²": "r2", "My RMSE": "rmse"})
+
+    check_display_structure(
+        display,
+        expected_metrics={"My R²", "My RMSE"},
+        expected_estimator_name="LinearRegression",
+    )
+
+
+def test_invalid_metric_type(linear_regression_with_test):
+    """An integer in the metric list raises a clear ValueError."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    err_msg = re.escape("Invalid type of metric: <class 'int'> for 1")
+    with pytest.raises(ValueError, match=err_msg):
+        report.metrics.summarize(metric=[1])
 
 
 def test_binary_classification_svc(svc_binary_classification_with_test):
-    """
-    Check the behaviour of summarize() with binary classification using SVC
-    (no predict_proba).
-    """
+    """SVC (no predict_proba): no ROC AUC, Log loss, or Brier score."""
     estimator, X_test, y_test = svc_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test, pos_label=1)
     display = report.metrics.summarize()
@@ -91,10 +160,7 @@ def test_binary_classification_svc(svc_binary_classification_with_test):
 
 
 def test_multiclass_classification_forest(forest_multiclass_classification_with_test):
-    """
-    Check the behaviour of summarize() with multiclass classification using
-    RandomForestClassifier.
-    """
+    """Multiclass classification with RandomForestClassifier."""
     estimator, X_test, y_test = forest_multiclass_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     display = report.metrics.summarize()
@@ -121,7 +187,7 @@ def test_multiclass_classification_forest(forest_multiclass_classification_with_
 
 
 def test_multiclass_classification_svc(svc_multiclass_classification_with_test):
-    """Check the behaviour of summarize() with multiclass classification using SVC."""
+    """Multiclass classification with SVC (no predict_proba)."""
     estimator, X_test, y_test = svc_multiclass_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     display = report.metrics.summarize()
@@ -146,7 +212,7 @@ def test_multiclass_classification_svc(svc_multiclass_classification_with_test):
 
 
 def test_regression(linear_regression_with_test):
-    """Check the behaviour of summarize() with regression."""
+    """Regression with LinearRegression."""
     estimator, X_test, y_test = linear_regression_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     display = report.metrics.summarize()
@@ -162,7 +228,7 @@ def test_regression(linear_regression_with_test):
 
 
 def test_multioutput_regression(linear_regression_multioutput_with_test):
-    """Check the behaviour of summarize() with multioutput regression."""
+    """Multioutput regression with LinearRegression."""
     estimator, X_test, y_test = linear_regression_multioutput_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     display = report.metrics.summarize(metric_kwargs={"multioutput": "raw_values"})
@@ -238,14 +304,12 @@ def test_data_source_both(forest_binary_classification_data):
 
 
 def test_unknown_ml_task(forest_binary_classification_with_test):
-    """Test summarize with unknown ML task."""
+    """Unknown ML task falls back to custom metric only."""
     estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
     report._ml_task = "unknown-task"
 
-    # If ML task is not recognized then none of the default metrics will
-    # work
     def custom_metric(y_true, y_pred):
         return 0.8
 
@@ -260,39 +324,88 @@ def test_unknown_ml_task(forest_binary_classification_with_test):
     assert display.data["output"].isna().all()
 
 
-# Test passing `metric`
+# ---------------------------------------------------------------------------
+# Metric strings
+# ---------------------------------------------------------------------------
 
 
-@pytest.mark.parametrize(
-    "metric, metric_kwargs",
-    [
-        ("accuracy", None),
-        ("neg_log_loss", None),
-        (accuracy_score, {"response_method": "predict"}),
-        (get_scorer("accuracy"), None),
-    ],
-)
-def test_metric_single_list_equivalence(
-    forest_binary_classification_with_test, metric, metric_kwargs
-):
-    """Check that passing a single string, callable, scorer is equivalent to passing a
-    list with a single element."""
+def test_metric_strings_plain(linear_regression_with_test):
+    """A list of skore built-in metric strings resolves to correct display names."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    display = report.metrics.summarize(metric=["r2", "rmse"])
+
+    check_display_structure(
+        display,
+        expected_metrics={"R²", "RMSE"},
+    )
+
+
+def test_metric_strings_with_extra_args(forest_binary_classification_with_test):
+    """Built-in metric strings with metric_kwargs change how the metric is computed.
+
+    Passing average='macro' collapses per-class precision/recall to a single row.
+    """
     estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-    display_single = report.metrics.summarize(
-        metric=metric, metric_kwargs=metric_kwargs
+
+    display = report.metrics.summarize(
+        metric=["precision", "recall"],
+        metric_kwargs={"average": "macro"},
     )
-    display_list = report.metrics.summarize(
-        metric=[metric], metric_kwargs=metric_kwargs
+
+    check_display_structure(
+        display,
+        expected_metrics={"Precision", "Recall"},
     )
-    pd.testing.assert_frame_equal(display_single.data, display_list.data)
+    assert (display.data["average"] == "macro").all()
 
 
-# Test passing sklearn metric strings
+def test_metric_strings_as_dict(linear_regression_with_test):
+    """Metric strings in dict form: dict keys override display names."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    display = report.metrics.summarize(
+        metric={
+            "Custom MAE": "neg_mean_absolute_error",
+            "Custom MSE": "neg_mean_squared_error",
+        }
+    )
+
+    check_display_structure(
+        display,
+        expected_metrics={"Custom MAE", "Custom MSE"},
+    )
+
+
+@pytest.mark.parametrize("metric", ["public_metric", "_private_metric"])
+def test_error_metric_strings(linear_regression_with_test, metric):
+    """An unrecognised metric string raises a clear ValueError."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    err_msg = re.escape(f"Invalid metric: {metric!r}.")
+    with pytest.raises(ValueError, match=err_msg):
+        report.metrics.summarize(metric=[metric])
+
+
+def test_sklearn_scorer_names_metric_kwargs(forest_binary_classification_with_test):
+    """metric_kwargs is not supported when metric is a sklearn scorer name string."""
+    classifier, X_test, y_test = forest_binary_classification_with_test
+    report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
+
+    err_msg = (
+        "The `metric_kwargs` parameter is not supported when `metric` is a "
+        "scikit-learn scorer name."
+    )
+    with pytest.raises(ValueError, match=err_msg):
+        report.metrics.summarize(metric=["f1"], metric_kwargs={"average": "macro"})
 
 
 def test_sklearn_metric_strings(forest_binary_classification_with_test):
-    """Check that multiple scikit-learn metric strings can be passed to summarize."""
+    """Multiple scikit-learn metric strings can be passed to summarize."""
     classifier, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
 
@@ -301,7 +414,7 @@ def test_sklearn_metric_strings(forest_binary_classification_with_test):
 
 
 def test_sklearn_metric_strings_regression(linear_regression_with_test):
-    """Test scikit-learn regression metric strings in summarize()."""
+    """Scikit-learn regression metric strings in summarize()."""
     regressor, X_test, y_test = linear_regression_with_test
     reg_report = EstimatorReport(regressor, X_test=X_test, y_test=y_test)
 
@@ -318,21 +431,106 @@ def test_sklearn_metric_strings_regression(linear_regression_with_test):
 
 
 def test_neg_metric_strings(forest_binary_classification_with_test):
-    """Check that scikit-learn metrics with 'neg_' prefix are handled correctly."""
+    """Scikit-learn metrics with 'neg_' prefix are handled correctly."""
     classifier, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(classifier, X_test=X_test, y_test=y_test)
 
     display = report.metrics.summarize(metric=["neg_log_loss"])
     assert isinstance(display.data, pd.DataFrame)
 
-    # Note: neg_log_loss was converted to log_loss
     assert "Log Loss" in set(display.data["metric"])
 
     score = display.data.set_index("metric").loc["Log Loss", "score"]
     assert score == pytest.approx(report.metrics.log_loss())
 
 
-# Test passing Scorers
+# ---------------------------------------------------------------------------
+# Metric scorers
+# ---------------------------------------------------------------------------
+
+
+def test_metric_scorers_plain(linear_regression_with_test):
+    """A list of sklearn scorers produces the correct metric names and scores."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    r2_scorer = make_scorer(r2_score, response_method="predict")
+    mae_scorer = make_scorer(
+        mean_absolute_error, response_method="predict", greater_is_better=False
+    )
+
+    display = report.metrics.summarize(metric=[r2_scorer, mae_scorer])
+
+    check_display_structure(
+        display, expected_metrics={"R2 Score", "Mean Absolute Error"}
+    )
+    scores = display.data.set_index("metric")["score"]
+    assert scores["R2 Score"] == pytest.approx(
+        r2_score(y_test, estimator.predict(X_test))
+    )
+    assert scores["Mean Absolute Error"] == pytest.approx(
+        mean_absolute_error(y_test, estimator.predict(X_test))
+    )
+
+
+def test_metric_scorers_with_extra_args(forest_binary_classification_with_test):
+    """Scorers created with make_scorer embed their own extra kwargs (e.g. average)."""
+    estimator, X_test, y_test = forest_binary_classification_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    f1_scorer = make_scorer(f1_score, response_method="predict", average="macro")
+    precision_scorer = make_scorer(
+        precision_score,
+        response_method="predict",
+        average="weighted",
+        zero_division=0,
+    )
+
+    display = report.metrics.summarize(metric=[f1_scorer, precision_scorer])
+
+    check_display_structure(
+        display,
+        expected_metrics={"F1 Score", "Precision Score"},
+    )
+    averages = display.data.set_index("metric")["average"]
+    assert averages["F1 Score"] == "macro"
+    assert averages["Precision Score"] == "weighted"
+
+
+def test_metric_scorers_as_dict(linear_regression_with_test):
+    """Scorers in dict form: dict keys become the display names."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    r2_scorer = make_scorer(r2_score, response_method="predict")
+    mae_scorer = make_scorer(
+        mean_absolute_error, response_method="predict", greater_is_better=False
+    )
+
+    display = report.metrics.summarize(
+        metric={"Custom R2": r2_scorer, "Custom MAE": mae_scorer}
+    )
+
+    check_display_structure(
+        display,
+        expected_metrics={"Custom R2", "Custom MAE"},
+    )
+
+
+def test_pos_label_scorer_error(forest_binary_classification_with_test):
+    """pos_label specified both in the scorer and in the report raises ValueError."""
+    estimator, X_test, y_test = forest_binary_classification_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test, pos_label=0)
+
+    f1_scorer = make_scorer(
+        f1_score, response_method="predict", average="macro", pos_label=1
+    )
+    err_msg = re.escape(
+        "The `pos_label` passed in the scorer and the one used when creating the "
+        "report must match; got 1 and 0."
+    )
+    with pytest.raises(ValueError, match=err_msg):
+        report.metrics.summarize(metric=[f1_scorer])
 
 
 @pytest.mark.parametrize(
@@ -350,8 +548,7 @@ def test_neg_metric_strings(forest_binary_classification_with_test):
 def test_scorer_binary_classification(
     forest_binary_classification_with_test, scorer, pos_label
 ):
-    """Check that we can pass scikit-learn scorer with different parameters to
-    summarize()."""
+    """Scorers with different pos_label configurations in binary classification."""
     estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(
         estimator, X_test=X_test, y_test=y_test, pos_label=pos_label
@@ -379,7 +576,7 @@ def test_scorer_binary_classification(
 def test_scorer_with_average(
     forest_multiclass_classification_with_test,
 ):
-    """Test multiclass classification with average parameter."""
+    """Multiclass classification with average parameter."""
     estimator, X_test, y_test = forest_multiclass_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
@@ -392,10 +589,7 @@ def test_scorer_with_average(
 
 
 def test_scorer_response_method_not_required_in_summarize(linear_regression_with_test):
-    """
-    When passing a scorer to summarize(), the
-    response_method embedded in the scorer should be used automatically
-    (the user should not need to pass it a second time).
+    """response_method embedded in the scorer is used automatically.
 
     Regression test for #2203.
     """
@@ -409,7 +603,6 @@ def test_scorer_response_method_not_required_in_summarize(linear_regression_with
         business_loss, greater_is_better=False, response_method="predict"
     )
 
-    # response_method is already in the scorer — no need to pass it explicitly
     display = report.metrics.summarize(metric=scorer)
 
     assert len(display.data) == 1
@@ -417,24 +610,97 @@ def test_scorer_response_method_not_required_in_summarize(linear_regression_with
     assert display.data["score"].iloc[0] == pytest.approx(expected)
 
 
-# Test passing `metric_kwargs`
+# ---------------------------------------------------------------------------
+# Metric callables
+# ---------------------------------------------------------------------------
 
 
-def test_metric_kwargs_average(forest_multiclass_classification_with_test):
-    """Check the behaviour of summarize() when `average` is passed in metric_kwargs."""
-    estimator, X_test, y_test = forest_multiclass_classification_with_test
+def test_metric_callables_plain(linear_regression_with_test):
+    """A list of plain callables requires response_method in metric_kwargs."""
+    estimator, X_test, y_test = linear_regression_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
-    display = report.metrics.summarize(metric_kwargs={"average": None})
 
-    assert isinstance(display.data, pd.DataFrame)
-    assert len(display.data) > 4
+    def my_mae(y_true, y_pred):
+        return mean_absolute_error(y_true, y_pred)
+
+    display = report.metrics.summarize(
+        metric=[my_mae],
+        metric_kwargs={"response_method": "predict"},
+    )
+
+    check_display_structure(
+        display,
+        expected_metrics={"My Mae"},
+    )
+    assert set(display.data["favorability"]) == {""}
+    score = display.data["score"].values[0]
+    assert score == pytest.approx(
+        mean_absolute_error(y_test, estimator.predict(X_test))
+    )
+
+
+def test_metric_callables_with_extra_args(linear_regression_with_test):
+    """Callables that need extra keyword arguments receive them through metric_kwargs."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    weights = np.ones_like(y_test) * 3.0
+
+    def weighted_mae(y_true, y_pred, weights):
+        return float(np.average(np.abs(y_true - y_pred), weights=weights))
+
+    display = report.metrics.summarize(
+        metric=[weighted_mae],
+        metric_kwargs={"response_method": "predict", "weights": weights},
+    )
+
+    check_display_structure(
+        display,
+        expected_metrics={"Weighted Mae"},
+    )
+    score = display.data["score"].values[0]
+    assert score == pytest.approx(
+        weighted_mae(y_test, estimator.predict(X_test), weights)
+    )
+
+
+def test_metric_callables_as_dict(linear_regression_with_test):
+    """Callables in dict form: dict keys become the display names."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    def my_metric(y_true, y_pred):
+        return mean_absolute_error(y_true, y_pred)
+
+    display = report.metrics.summarize(
+        metric={"My Custom Error": my_metric},
+        metric_kwargs={"response_method": "predict"},
+    )
+
+    check_display_structure(
+        display,
+        expected_metrics={"My Custom Error"},
+    )
+    score = display.data["score"].values[0]
+    assert score == pytest.approx(
+        mean_absolute_error(y_test, estimator.predict(X_test))
+    )
+
+
+def test_custom_metric_no_response_method(forest_binary_classification_with_test):
+    """A callable without response_method raises a descriptive ValueError."""
+    estimator, X_test, y_test = forest_binary_classification_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    def custom_metric(y_true, y_pred):
+        return 0.5
+
+    with pytest.raises(ValueError, match="response_method is required"):
+        report.metrics.summarize(metric=custom_metric)
 
 
 def test_custom_metric_average_none(forest_binary_classification_with_test):
-    """
-    Check that passing arguments to a custom metric through metric_kwargs
-    works correctly.
-    """
+    """Passing arguments to a custom metric through metric_kwargs works correctly."""
     estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
@@ -450,9 +716,126 @@ def test_custom_metric_average_none(forest_binary_classification_with_test):
     assert set(display.data["label"]) == {0, 1}
 
 
+# ---------------------------------------------------------------------------
+# Mixed metric types
+# --------------------------------------------------------------------------
+
+
+@pytest.mark.parametrize(
+    "metric, metric_kwargs",
+    [
+        ("accuracy", None),
+        ("neg_log_loss", None),
+        (accuracy_score, {"response_method": "predict"}),
+        (get_scorer("accuracy"), None),
+    ],
+)
+def test_metric_single_list_equivalence(
+    forest_binary_classification_with_test, metric, metric_kwargs
+):
+    """Passing a single metric is equivalent to passing a list with one element."""
+    estimator, X_test, y_test = forest_binary_classification_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+    display_single = report.metrics.summarize(
+        metric=metric, metric_kwargs=metric_kwargs
+    )
+    display_list = report.metrics.summarize(
+        metric=[metric], metric_kwargs=metric_kwargs
+    )
+    pd.testing.assert_frame_equal(display_single.data, display_list.data)
+
+
+def test_metric_mix_plain(linear_regression_with_test):
+    """A list that mixes built-in string, scorer, and callable in one call."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    mae_scorer = make_scorer(
+        mean_absolute_error, response_method="predict", greater_is_better=False
+    )
+
+    def my_r2(y_true, y_pred):
+        return r2_score(y_true, y_pred)
+
+    display = report.metrics.summarize(
+        metric=["rmse", mae_scorer, my_r2],
+        metric_kwargs={"response_method": "predict"},
+    )
+
+    check_display_structure(
+        display,
+        expected_metrics={"RMSE", "Mean Absolute Error", "My R2"},
+    )
+
+
+def test_metric_mix_with_extra_args(linear_regression_with_test):
+    """A mixed list where the callable needs extra kwargs from metric_kwargs."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    mae_scorer = make_scorer(
+        mean_absolute_error, response_method="predict", greater_is_better=False
+    )
+
+    def scaled_mae(y_true, y_pred, scale=1.0):
+        return scale * mean_absolute_error(y_true, y_pred)
+
+    display = report.metrics.summarize(
+        metric=["r2", mae_scorer, scaled_mae],
+        metric_kwargs={"response_method": "predict", "scale": 2.0},
+    )
+
+    check_display_structure(
+        display,
+        expected_metrics={"R²", "Mean Absolute Error", "Scaled Mae"},
+    )
+    scores = display.data.set_index("metric")["score"]
+    plain_mae = mean_absolute_error(y_test, estimator.predict(X_test))
+    assert scores["Scaled Mae"] == pytest.approx(2.0 * plain_mae)
+
+
+def test_metric_mix_as_dict(linear_regression_with_test):
+    """A dict mixing built-in string, scorer, and callable with custom display names."""
+    estimator, X_test, y_test = linear_regression_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    r2_scorer = make_scorer(r2_score, response_method="predict")
+
+    def my_mae(y_true, y_pred):
+        return mean_absolute_error(y_true, y_pred)
+
+    display = report.metrics.summarize(
+        metric={
+            "Skore RMSE": "rmse",
+            "Scorer R2": r2_scorer,
+            "Callable MAE": my_mae,
+        },
+        metric_kwargs={"response_method": "predict"},
+    )
+
+    check_display_structure(
+        display,
+        expected_metrics={"Skore RMSE", "Scorer R2", "Callable MAE"},
+    )
+
+
+# ---------------------------------------------------------------------------
+# metric_kwargs
+# ---------------------------------------------------------------------------
+
+
+def test_metric_kwargs_average(forest_multiclass_classification_with_test):
+    """average passed in metric_kwargs expands per-class metrics."""
+    estimator, X_test, y_test = forest_multiclass_classification_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+    display = report.metrics.summarize(metric_kwargs={"average": None})
+
+    assert isinstance(display.data, pd.DataFrame)
+    assert len(display.data) > 4
+
+
 def test_metric_kwargs_multioutput(linear_regression_multioutput_with_test):
-    """Check the behaviour of summarize() when `multioutput` is passed in
-    metric_kwargs."""
+    """multioutput passed in metric_kwargs produces per-output rows."""
     estimator, X_test, y_test = linear_regression_multioutput_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
     display = report.metrics.summarize(metric_kwargs={"multioutput": "raw_values"})
@@ -463,7 +846,7 @@ def test_metric_kwargs_multioutput(linear_regression_multioutput_with_test):
 
 
 def test_metric_kwargs_none(forest_binary_classification_with_test):
-    """Test callable metric when metric_kwargs is None."""
+    """Callable metric when metric_kwargs is None."""
     estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
 
@@ -478,9 +861,7 @@ def test_metric_kwargs_none(forest_binary_classification_with_test):
 
 
 def test_metric_kwargs_override_scorer(linear_regression_with_test):
-    """
-    When a scorer has kwargs baked in,
-    passing metric_kwargs to summarize() should override them.
+    """metric_kwargs overrides kwargs baked into the scorer.
 
     Regression test for #2203 follow-up.
     """
@@ -507,11 +888,13 @@ def test_metric_kwargs_override_scorer(linear_regression_with_test):
     assert display.data["score"].iloc[0] == pytest.approx(expected)
 
 
-# Test passing `pos_label`
+# ---------------------------------------------------------------------------
+# pos_label
+# ---------------------------------------------------------------------------
 
 
 def test_pos_label(forest_binary_classification_with_test):
-    """Check that `pos_label` can be passed."""
+    """pos_label collapses per-class metrics to a single row."""
     estimator, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(estimator, X_test=X_test, y_test=y_test, pos_label=1)
     display = report.metrics.summarize()
@@ -533,15 +916,12 @@ def test_pos_label(forest_binary_classification_with_test):
 
     assert len(display.data[display.data["metric"] == "Precision"]) == 1
     assert len(display.data[display.data["metric"] == "Recall"]) == 1
-    # Label is None for precision/recall when pos_label is specified
     assert display.data["label"].isna().all()
     assert display.data["output"].isna().all()
 
 
 def test_pos_label_strings(forest_binary_classification_with_test):
-    """
-    Check the behaviour of summarize() with binary classification using string labels.
-    """
+    """Binary classification with string labels."""
     estimator, X_test, y_test = forest_binary_classification_with_test
 
     target_names = np.array(["neg", "pos"], dtype=object)
@@ -568,9 +948,7 @@ def test_pos_label_strings(forest_binary_classification_with_test):
 
 
 def test_pos_label_bool(forest_binary_classification_with_test):
-    """
-    Check the behaviour of summarize() with binary classification using boolean labels.
-    """
+    """Binary classification with boolean labels."""
     estimator, X_test, y_test = forest_binary_classification_with_test
 
     target_names = np.array([False, True], dtype=bool)
@@ -593,7 +971,6 @@ def test_pos_label_bool(forest_binary_classification_with_test):
     }
 
     labels = display.data.set_index("metric").loc["Precision", "label"]
-    # Use `is` to avoid casting
     assert any(label is np.False_ for label in labels)
     assert any(label is np.True_ for label in labels)
 
@@ -601,7 +978,7 @@ def test_pos_label_bool(forest_binary_classification_with_test):
 def test_pos_label_scorer_names(
     forest_binary_classification_with_test,
 ):
-    """Check that `pos_label` is dispatched with scikit-learn scorer names."""
+    """pos_label is dispatched with scikit-learn scorer names."""
     classifier, X_test, y_test = forest_binary_classification_with_test
     report = EstimatorReport(classifier, X_test=X_test, y_test=y_test, pos_label=0)
 
@@ -620,7 +997,7 @@ def test_pos_label_scorer_names(
     "metric, metric_fn", [("precision", precision_score), ("recall", recall_score)]
 )
 def test_pos_label_overwrite(metric, metric_fn):
-    """Check that `pos_label` can be set when creating the report."""
+    """pos_label can be set when creating the report."""
     X, y = make_classification(
         n_classes=2, class_sep=0.8, weights=[0.4, 0.6], random_state=0
     )
@@ -628,23 +1005,64 @@ def test_pos_label_overwrite(metric, metric_fn):
     y = labels[y]
     classifier = LogisticRegression().fit(X, y)
 
-    # Test without pos_label - should have multiple rows (one per class)
+    # Without pos_label - should have multiple rows (one per class)
     report = EstimatorReport(classifier, X_test=X, y_test=y)
     display = report.metrics.summarize(metric=metric)
     assert isinstance(display.data, pd.DataFrame)
-    assert len(display.data) == 2  # One row per class
+    assert len(display.data) == 2
     assert set(display.data["label"]) == {"A", "B"}
 
-    # Test with pos_label="B" - should have single row
+    # With pos_label="B" - should have single row
     report = EstimatorReport(classifier, X_test=X, y_test=y, pos_label="B")
     display = report.metrics.summarize(metric=metric)
     assert len(display.data) == 1
     score_B = display.data["score"].values[0]
     assert score_B == pytest.approx(metric_fn(y, classifier.predict(X), pos_label="B"))
 
-    # Test with pos_label="A" - should have single row
+    # With pos_label="A" - should have single row
     report = EstimatorReport(classifier, X_test=X, y_test=y, pos_label="A")
     display = report.metrics.summarize(metric=metric)
     assert len(display.data) == 1
     score_A = display.data["score"].values[0]
     assert score_A == pytest.approx(metric_fn(y, classifier.predict(X), pos_label="A"))
+
+
+# ---------------------------------------------------------------------------
+# Cache and data_source
+# ---------------------------------------------------------------------------
+
+
+def test_cache(forest_binary_classification_with_test):
+    """summarize() results are cached; second call returns the same data."""
+    estimator, X_test, y_test = forest_binary_classification_with_test
+    report = EstimatorReport(estimator, X_test=X_test, y_test=y_test)
+
+    with check_cache_changed(report._cache):
+        result = report.metrics.summarize()
+    assert isinstance(result, MetricsSummaryDisplay)
+
+    with check_cache_unchanged(report._cache):
+        result_from_cache = report.metrics.summarize()
+    assert_frame_equal(result.data, result_from_cache.data)
+
+
+def test_data_source_both(forest_binary_classification_data):
+    """data_source='both' concatenates train and test results."""
+    estimator, X, y = forest_binary_classification_data
+    X_train, X_test, y_train, y_test = train_test_split(X, y, random_state=42)
+
+    report = EstimatorReport(
+        estimator, X_train=X_train, y_train=y_train, X_test=X_test, y_test=y_test
+    )
+
+    display_train = report.metrics.summarize(data_source="train")
+    display_test = report.metrics.summarize(data_source="test")
+    display_both = report.metrics.summarize(data_source="both")
+
+    assert set(display_both.data["data_source"]) == {"train", "test"}
+
+    train_data = display_both.data[display_both.data["data_source"] == "train"]
+    assert_array_equal(train_data["score"], display_train.data["score"])
+
+    test_data = display_both.data[display_both.data["data_source"] == "test"]
+    assert_array_equal(test_data["score"], display_test.data["score"])
